Refactor the whole thing
The file .bashrc_synced file gets sourced from the original .profile
file, and all other .*_synced files are sourced through it.

Files not ending with _synced have to be either linked or sourced
in their main counterparts contained in the home directory.

The script dotinit.sh got updated so that it solves some of this.